add function increment_chest_bar
I wrote this small function for incrementing/decrementing the inventory limit (a.k.a. bar) of a chest. I don't know the event handling and key pressing side of the code so I simply made a function that takes the chest entity as a parameter, along with a bool to indicate whether to increase/decrease, and the amount to increase/decrease by with a default of 1. For example, you can make the code call this function upon the player pressing page_up or page_down on the chest inventory menu in order to increase/decrease the limit by 1.